Remove abstract claimed returned resource CIRC-742
